requirements.txt
lyricsgenius
numpy
spotipy
python-dotenv
torch
transformers
pinecone